Revert past six commits
This patch reverts the past six commits as they were accidentally pushed
to main.
